requirements.txt
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0
scikit_learn==1.7.2
scipy==1.16.3
skimage==0.25.2
